fix(delivery,console): the review pass — a poll that never stopped, and comments that argued the journey
Seventeen findings from /code-review over 7a40ad58, across both axes. One is a
real regression I introduced; the rest are documentation that argued for the
change instead of describing what shipped.

**A POLL THAT NEVER STOPPED.** `statusIsMoving` gained a `none` clause scoped by
`build.status === "succeeded" && deploy.status === "deployed"`, on the theory
that the build state screens off versions nothing will ever judge. It does not.
`stateForUnlandedValidation` settles quota-blocked, publisher-credentials,
conflict-budget and redispatch-budget WITHOUT recording a verdict, and
`settleJudged` closes the version's validation task on the way out — so nothing
restarts it and `validationStageFromRun` answers `none` for good. `build.status`
cannot see any of that: it is derived from the newest DEV run, which succeeded.
All three clauses therefore stay true forever, on a hook every route mounts for
the toolbar badge, fanning out to four backends. Org-quota exhaustion during
validation is an ordinary way to get there.

The clause is removed rather than re-scoped, because no clause over this payload
can tell the two `none`s apart. #649 asked for the pending state on the fast
cadence, and its own cost constraint — "only while a verdict is outstanding" —
is what the clause violated; honouring the constraint means dropping it. The
wait it was buying is answered where it is caused: the settling run's reconcile
makes the gap seconds rather than a sweep interval, and what remains is one idle
poll before "Validating" appears. `running` and `awaiting-fix` stay, so the
verdict still lands within 5s instead of 30s. Splitting `none` properly is a
contract change and #647's deferred item.

The predicate moves to `features/projects/lib/statusCadence.ts` beside the other
pure predicates, which retires the "exported for its own test" apology; `api/`
already imports two such helpers. Its test moves with it — and gained the
repo-rung and spec-interview cases, because moving the function silently dropped
`repoStatus === "pending" | "cloning"` and nothing failed. Restored, and now
pinned.

**COMMENTS THAT ARGUED THE JOURNEY.** `services/AGENTS.md` requires docs to
describe the shipped end state. Five places narrated the before-picture instead
("the sweep WAS not a backstop", "a version SAT unjudged", "that residual GOES
when none is split") — and `milestone_run.go` had drifted into flatly
contradicting `sweep.go` in the same commit: one said the sweep is not a
backstop but a driver, the other that it is a backstop and not a driver. All
rewritten in the present tense, and the "three hand-offs / no webhook / #649"
paragraph now lives once, on `SettleHandsWorkOnward`, with pointers from the six
places that had restated it.

**A JUSTIFICATION THAT WAS FALSE.** `MilestoneReconciler`'s doc claimed a shared
struct would force the two packages to import each other. `StartRunRequest`
disproves it — a type at the domain ROOT with a narrow port in each package is
exactly the documented answer to `slice ⊥ sibling`. The signature stands, but
for the real reason: the supervisor's three other ports onto this package all
pass the milestone as an int, and a root type earns its place by carrying a
domain concept several callers compose, where this one carries an address.

**THE `cancelled` EXCLUSION WAS OVERSOLD.** It claimed to honour the person who
cancelled. It cannot: `milestoneCancelled` suppresses dev alone, and a validation
run cancelled before its first read leaves the task open, so the next sweep pass
restarts the judging regardless. What the exclusion actually buys is that the
platform does not contradict the click within the same second. Now documented as
the courtesy it is. `blocked` is unchanged — that one is a genuine spin.

**THE ORIGIN NOTE, DECLINED WITH A REASON.** #649 asks that `startValidationRun`
stop stamping `revalidate` on a version's first judgement, "corrected alongside,
or filed separately". Neither, and the premise is why: the note says the console
labels such a run a revalidation, and it does not. The console reads `origin` at
exactly one site, as the kind fallback for rows predating the kind column
(`runView.ts` KIND_FOR_ORIGIN); nothing renders it. A fourth value would add a
member to a closed contract enum, a legacy mapping that can never be hit, and
fixtures, to separate two triggers no surface tells apart. Recorded in
`adopt.go` beside the choice, with the condition that would change the answer.

Also: the dead `e.p.Issues == nil` guard in `reconcileRepo` (ReconcileMilestone
re-checks it), `omitempty` dropped from `MilestoneTitle` to match its sibling
inputs, and the speculative "a caller may pass \"\"" removed — neither caller does.

Gates: go build/vet 0, go test ./... exit 0 (69 packages), deadcode OK, gofmt
clean, console typecheck/lint 0, 137 files / 1733 tests. The re-added `none`
clause fails the two tests that now cover it.
